+++ tests/test_flatland_malfunction.py
@@ -66,8 +66,7 @@ class SingleAgentNavigationObs(ObservationBuilder):
 
 def test_malfunction_process():
     # Set fixed malfunction duration for this test
-    stochastic_data = {'prop_malfunction': 1.,
-                       'malfunction_rate': 1000,
+    stochastic_data = {'malfunction_rate': 1,
                        'min_duration': 3,
                        'max_duration': 3}
 
@@ -84,11 +83,6 @@ def test_malfunction_process():
     # reset to initialize agents_static
     obs, info = env.reset(False, False, True, random_seed=10)
 
-    # Check that a initial duration for malfunction was assigned
-    assert env.agents[0].malfunction_data['next_malfunction'] > 0
-    for agent in env.agents:
-        agent.status = RailAgentStatus.ACTIVE
-
     agent_halts = 0
     total_down_time = 0
     agent_old_position = env.agents[0].position
@@ -100,12 +94,6 @@ def test_malfunction_process():
 
         for i in range(len(obs)):
             actions[i] = np.argmax(obs[i]) + 1
-
-        if step % 5 == 0:
-            # Stop the agent and set it to be malfunctioning
-            env.agents[0].malfunction_data['malfunction'] = -1
-            env.agents[0].malfunction_data['next_malfunction'] = 0
-            agent_halts += 1
 
         obs, all_rewards, done, _ = env.step(actions)
 
@@ -122,11 +110,8 @@ def test_malfunction_process():
         total_down_time += env.agents[0].malfunction_data['malfunction']
 
     # Check that the appropriate number of malfunctions is achieved
-    assert env.agents[0].malfunction_data['nr_malfunctions'] == 20, "Actual {}".format(
+    assert env.agents[0].malfunction_data['nr_malfunctions'] == 23, "Actual {}".format(
         env.agents[0].malfunction_data['nr_malfunctions'])
-
-    # Check that 20 stops where performed
-    assert agent_halts == 20
 
     # Check that malfunctioning data was standing around
     assert total_down_time > 0
@@ -135,8 +120,7 @@ def test_malfunction_process():
 def test_malfunction_process_statistically():
     """Tests hat malfunctions are produced by stochastic_data!"""
     # Set fixed malfunction duration for this test
-    stochastic_data = {'prop_malfunction': 1.,
-                       'malfunction_rate': 5,
+    stochastic_data = {'malfunction_rate': 5,
                        'min_duration': 5,
                        'max_duration': 5}
 
@@ -155,17 +139,18 @@ def test_malfunction_process_statistically():
     env.reset(True, True, False, random_seed=10)
 
     env.agents[0].target = (0, 0)
-
-    agent_malfunction_list = [[0, 0, 0, 0, 0, 0, 6, 5, 4, 3, 2, 1, 0, 6, 5, 4, 3, 2, 1, 0],
-                              [0, 0, 0, 0, 0, 0, 0, 6, 5, 4, 3, 2, 1, 0, 0, 0, 0, 0, 6, 5],
-                              [0, 0, 0, 0, 0, 0, 0, 0, 0, 6, 5, 4, 3, 2, 1, 0, 0, 6, 5, 4],
-                              [0, 0, 0, 0, 0, 6, 5, 4, 3, 2, 1, 0, 0, 0, 0, 0, 0, 6, 5, 4],
-                              [6, 6, 5, 4, 3, 2, 1, 0, 6, 5, 4, 3, 2, 1, 0, 0, 0, 0, 0, 0],
-                              [6, 6, 5, 4, 3, 2, 1, 0, 6, 5, 4, 3, 2, 1, 0, 0, 6, 5, 4, 3],
-                              [0, 0, 0, 0, 6, 5, 4, 3, 2, 1, 0, 6, 5, 4, 3, 2, 1, 0, 6, 5],
-                              [0, 0, 0, 0, 0, 6, 5, 4, 3, 2, 1, 0, 0, 6, 5, 4, 3, 2, 1, 0],
-                              [0, 0, 0, 0, 0, 0, 0, 6, 5, 4, 3, 2, 1, 0, 6, 5, 4, 3, 2, 1],
-                              [6, 6, 6, 6, 5, 4, 3, 2, 1, 0, 0, 0, 0, 6, 5, 4, 3, 2, 1, 0]]
+    # Next line only for test generation
+    #agent_malfunction_list = [[] for i in range(10)]
+    agent_malfunction_list = [[0, 5, 4, 3, 2, 1, 0, 0, 0, 5, 4, 3, 2, 1, 0, 5, 4, 3, 2, 1],
+     [0, 0, 0, 0, 0, 0, 5, 4, 3, 2, 1, 0, 0, 5, 4, 3, 2, 1, 0, 0],
+     [5, 4, 3, 2, 1, 0, 5, 4, 3, 2, 1, 0, 0, 0, 5, 4, 3, 2, 1, 0],
+     [0, 5, 4, 3, 2, 1, 0, 0, 0, 5, 4, 3, 2, 1, 0, 5, 4, 3, 2, 1],
+     [0, 0, 0, 0, 0, 0, 0, 0, 5, 4, 3, 2, 1, 0, 0, 0, 0, 0, 0, 0],
+     [0, 0, 0, 0, 0, 0, 5, 4, 3, 2, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0],
+     [0, 0, 0, 0, 0, 5, 4, 3, 2, 1, 0, 0, 0, 0, 0, 5, 4, 3, 2, 1],
+     [0, 0, 0, 5, 4, 3, 2, 1, 0, 0, 5, 4, 3, 2, 1, 0, 5, 4, 3, 2],
+     [5, 4, 3, 2, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0],
+     [5, 4, 3, 2, 1, 0, 0, 0, 0, 5, 4, 3, 2, 1, 0, 0, 0, 0, 0, 0]]
 
     for step in range(20):
         action_dict: Dict[int, RailEnvActions] = {}
@@ -173,16 +158,16 @@ def test_malfunction_process_statistically():
             # We randomly select an action
             action_dict[agent_idx] = RailEnvActions(np.random.randint(4))
             # For generating tests only:
-            # agent_malfunction_list[agent_idx].append(env.agents[agent_idx].malfunction_data['malfunction'])
+            #agent_malfunction_list[agent_idx].append(env.agents[agent_idx].malfunction_data['malfunction'])
             assert env.agents[agent_idx].malfunction_data['malfunction'] == agent_malfunction_list[agent_idx][step]
         env.step(action_dict)
+    #print(agent_malfunction_list)
 
 
 def test_malfunction_before_entry():
-    """Tests that malfunctions are produced by stochastic_data!"""
+    """Tests that malfunctions are working properly for agents before entering the environment!"""
     # Set fixed malfunction duration for this test
-    stochastic_data = {'prop_malfunction': 1.,
-                       'malfunction_rate': 1,
+    stochastic_data = {'malfunction_rate': 2,
                        'min_duration': 10,
                        'max_duration': 10}
 
@@ -191,7 +176,7 @@ def test_malfunction_before_entry():
     env = RailEnv(width=25,
                   height=30,
                   rail_generator=rail_from_grid_transition_map(rail),
-                  schedule_generator=random_schedule_generator(seed=2),  # seed 12
+                  schedule_generator=random_schedule_generator(seed=1),  # seed 12
                   number_of_agents=10,
                   random_seed=1,
                   stochastic_data=stochastic_data,  # Malfunction data generator
@@ -200,46 +185,62 @@ def test_malfunction_before_entry():
     env.reset(False, False, False, random_seed=10)
     env.agents[0].target = (0, 0)
 
-    # Print for test generation
-    assert env.agents[0].malfunction_data['malfunction'] == 11
-    assert env.agents[1].malfunction_data['malfunction'] == 11
-    assert env.agents[2].malfunction_data['malfunction'] == 11
-    assert env.agents[3].malfunction_data['malfunction'] == 11
-    assert env.agents[4].malfunction_data['malfunction'] == 11
-    assert env.agents[5].malfunction_data['malfunction'] == 11
-    assert env.agents[6].malfunction_data['malfunction'] == 11
-    assert env.agents[7].malfunction_data['malfunction'] == 11
-    assert env.agents[8].malfunction_data['malfunction'] == 11
-    assert env.agents[9].malfunction_data['malfunction'] == 11
+    # Test initial malfunction values for all agents
+    # we want some agents to be malfuncitoning already and some to be working
+    # we want different next_malfunction values for the agents
+    assert env.agents[0].malfunction_data['malfunction'] == 0
+    assert env.agents[1].malfunction_data['malfunction'] == 0
+    assert env.agents[2].malfunction_data['malfunction'] == 10
+    assert env.agents[3].malfunction_data['malfunction'] == 0
+    assert env.agents[4].malfunction_data['malfunction'] == 0
+    assert env.agents[5].malfunction_data['malfunction'] == 0
+    assert env.agents[6].malfunction_data['malfunction'] == 0
+    assert env.agents[7].malfunction_data['malfunction'] == 0
+    assert env.agents[8].malfunction_data['malfunction'] == 10
+    assert env.agents[9].malfunction_data['malfunction'] == 10
 
-    for step in range(20):
-        action_dict: Dict[int, RailEnvActions] = {}
-        for agent in env.agents:
-            # We randomly select an action
-            action_dict[agent.handle] = RailEnvActions(2)
-            if step < 10:
-                action_dict[agent.handle] = RailEnvActions(0)
+    #for a in range(10):
+    #  print("assert env.agents[{}].malfunction_data['malfunction'] == {}".format(a,env.agents[a].malfunction_data['malfunction']))
 
+
+def test_malfunction_values_and_behavior():
+    """
+    Test the malfunction counts down as desired
+    Returns
+    -------
+
+    """
+    # Set fixed malfunction duration for this test
+
+    rail, rail_map = make_simple_rail2()
+    action_dict: Dict[int, RailEnvActions] = {}
+    stochastic_data = {'malfunction_rate': 0.001,
+                       'min_duration': 10,
+                       'max_duration': 10}
+    env = RailEnv(width=25,
+                  height=30,
+                  rail_generator=rail_from_grid_transition_map(rail),
+                  schedule_generator=random_schedule_generator(seed=2),  # seed 12
+                  stochastic_data=stochastic_data,
+                  number_of_agents=1,
+                  random_seed=1,
+                  )
+
+    # reset to initialize agents_static
+    env.reset(False, False, activate_agents=True, random_seed=10)
+
+    # Assertions
+    assert_list = [9, 8, 7, 6, 5, 4, 3, 2, 1, 0, 10, 9, 8, 7, 6, 5]
+    print("[")
+    for time_step in range(15):
+        # Move in the env
         env.step(action_dict)
-    assert env.agents[1].malfunction_data['malfunction'] == 2
-    assert env.agents[2].malfunction_data['malfunction'] == 2
-    assert env.agents[3].malfunction_data['malfunction'] == 2
-    assert env.agents[4].malfunction_data['malfunction'] == 2
-    assert env.agents[5].malfunction_data['malfunction'] == 2
-    assert env.agents[6].malfunction_data['malfunction'] == 2
-    assert env.agents[7].malfunction_data['malfunction'] == 2
-    assert env.agents[8].malfunction_data['malfunction'] == 2
-    assert env.agents[9].malfunction_data['malfunction'] == 2
-
-    # for a in range(env.get_num_agents()):
-    #    print("assert env.agents[{}].malfunction_data['malfunction'] == {}".format(a,
-    #                                                                               env.agents[a].malfunction_data[
-    #                                                                                   'malfunction']))
+        # Check that next_step decreases as expected
+        assert env.agents[0].malfunction_data['malfunction'] == assert_list[time_step]
 
 
 def test_initial_malfunction():
-    stochastic_data = {'prop_malfunction': 1.,  # Percentage of defective agents
-                       'malfunction_rate': 100,  # Rate of malfunction occurence
+    stochastic_data = {'malfunction_rate': 1000,  # Rate of malfunction occurence
                        'min_duration': 2,  # Minimal duration of malfunction
                        'max_duration': 5  # Max duration of malfunction
                        }
@@ -283,22 +284,22 @@ def test_initial_malfunction():
                 direction=Grid4TransitionsEnum.EAST,
                 action=RailEnvActions.MOVE_FORWARD,
                 malfunction=1,
-                reward=env.start_penalty + env.step_penalty * 1.0
-                # malfunctioning ends: starting and running at speed 1.0
+                reward=env.step_penalty
+
+            ),  # malfunctioning ends: starting and running at speed 1.0
+            Replay(
+                position=(3, 2),
+                direction=Grid4TransitionsEnum.EAST,
+                action=RailEnvActions.MOVE_FORWARD,
+                malfunction=0,
+                reward=env.start_penalty + env.step_penalty * 1.0  # running at speed 1.0
             ),
             Replay(
                 position=(3, 3),
                 direction=Grid4TransitionsEnum.EAST,
                 action=RailEnvActions.MOVE_FORWARD,
                 malfunction=0,
-                reward=env.step_penalty * 1.0  # running at speed 1.0
-            ),
-            Replay(
-                position=(3, 4),
-                direction=Grid4TransitionsEnum.EAST,
-                action=RailEnvActions.MOVE_FORWARD,
-                malfunction=0,
-                reward=env.step_penalty * 1.0  # running at speed 1.0
+                reward=env.step_penalty  # running at speed 1.0
             )
         ],
         speed=env.agents[0].speed_data['speed'],
@@ -346,7 +347,7 @@ def test_initial_malfunction_stop_moving():
                 position=(3, 2),
                 direction=Grid4TransitionsEnum.EAST,
                 action=RailEnvActions.DO_NOTHING,
-                malfunction=3,
+                malfunction=2,
                 reward=env.step_penalty,  # full step penalty when stopped
                 status=RailAgentStatus.ACTIVE
             ),
@@ -357,7 +358,7 @@ def test_initial_malfunction_stop_moving():
                 position=(3, 2),
                 direction=Grid4TransitionsEnum.EAST,
                 action=RailEnvActions.STOP_MOVING,
-                malfunction=2,
+                malfunction=1,
                 reward=env.step_penalty,  # full step penalty while stopped
                 status=RailAgentStatus.ACTIVE
             ),
@@ -366,7 +367,7 @@ def test_initial_malfunction_stop_moving():
                 position=(3, 2),
                 direction=Grid4TransitionsEnum.EAST,
                 action=RailEnvActions.DO_NOTHING,
-                malfunction=1,
+                malfunction=0,
                 reward=env.step_penalty,  # full step penalty while stopped
                 status=RailAgentStatus.ACTIVE
             ),
@@ -434,7 +435,7 @@ def test_initial_malfunction_do_nothing():
                 position=(3, 2),
                 direction=Grid4TransitionsEnum.EAST,
                 action=RailEnvActions.DO_NOTHING,
-                malfunction=3,
+                malfunction=2,
                 reward=env.step_penalty,  # full step penalty while malfunctioning
                 status=RailAgentStatus.ACTIVE
             ),
@@ -445,7 +446,7 @@ def test_initial_malfunction_do_nothing():
                 position=(3, 2),
                 direction=Grid4TransitionsEnum.EAST,
                 action=RailEnvActions.DO_NOTHING,
-                malfunction=2,
+                malfunction=1,
                 reward=env.step_penalty,  # full step penalty while stopped
                 status=RailAgentStatus.ACTIVE
             ),
@@ -454,7 +455,7 @@ def test_initial_malfunction_do_nothing():
                 position=(3, 2),
                 direction=Grid4TransitionsEnum.EAST,
                 action=RailEnvActions.DO_NOTHING,
-                malfunction=1,
+                malfunction=0,
                 reward=env.step_penalty,  # full step penalty while stopped
                 status=RailAgentStatus.ACTIVE
             ),
@@ -484,45 +485,14 @@ def test_initial_malfunction_do_nothing():
     run_replay_config(env, [replay_config], activate_agents=False)
 
 
-def test_initial_nextmalfunction_not_below_zero():
-    random.seed(0)
-    np.random.seed(0)
-
-    stochastic_data = {'prop_malfunction': 1.,  # Percentage of defective agents
-                       'malfunction_rate': 70,  # Rate of malfunction occurence
-                       'min_duration': 2,  # Minimal duration of malfunction
-                       'max_duration': 5  # Max duration of malfunction
-                       }
-
-    rail, rail_map = make_simple_rail2()
-
-    env = RailEnv(width=25,
-                  height=30,
-                  rail_generator=rail_from_grid_transition_map(rail),
-                  schedule_generator=random_schedule_generator(),
-                  number_of_agents=1,
-                  stochastic_data=stochastic_data,  # Malfunction data generator
-                  obs_builder_object=SingleAgentNavigationObs()
-                  )
-    # reset to initialize agents_static
-    env.reset()
-    agent = env.agents[0]
-    env.step({})
-    # was next_malfunction was -1 befor the bugfix https://gitlab.aicrowd.com/flatland/flatland/issues/186
-    assert agent.malfunction_data['next_malfunction'] >= 0, \
-        "next_malfunction should be >=0, found {}".format(agent.malfunction_data['next_malfunction'])
-
-
 def tests_random_interference_from_outside():
     """Tests that malfunctions are produced by stochastic_data!"""
     # Set fixed malfunction duration for this test
-    stochastic_data = {'prop_malfunction': 1.,
-                       'malfunction_rate': 1,
+    stochastic_data = {'malfunction_rate': 1,
                        'min_duration': 10,
                        'max_duration': 10}
 
     rail, rail_map = make_simple_rail2()
-
     env = RailEnv(width=25,
                   height=30,
                   rail_generator=rail_from_grid_transition_map(rail),
@@ -534,9 +504,7 @@ def tests_random_interference_from_outside():
     env.reset()
     # reset to initialize agents_static
     env.agents[0].speed_data['speed'] = 0.33
-    env.agents[0].initial_position = (3, 0)
-    env.agents[0].target = (3, 9)
-    env.reset(False, False, False)
+    env.reset(False, False, False, random_seed=10)
     env_data = []
 
     for step in range(200):
@@ -567,11 +535,8 @@ def tests_random_interference_from_outside():
     env.reset()
     # reset to initialize agents_static
     env.agents[0].speed_data['speed'] = 0.33
-    env.agents[0].initial_position = (3, 0)
-    env.agents[0].target = (3, 9)
-    env.reset(False, False, False)
+    env.reset(False, False, False, random_seed=10)
 
-    # Print for test generation
     dummy_list = [1, 2, 6, 7, 8, 9, 4, 5, 4]
     for step in range(200):
         action_dict: Dict[int, RailEnvActions] = {}
@@ -586,3 +551,57 @@ def tests_random_interference_from_outside():
         _, reward, _, _ = env.step(action_dict)
         assert reward[0] == env_data[step][0]
         assert env.agents[0].position == env_data[step][1]
+
+
+def test_last_malfunction_step():
+    """
+    Test to check that agent moves when it is not malfunctioning
+
+    """
+
+    # Set fixed malfunction duration for this test
+    stochastic_data = {'malfunction_rate': 5,
+                       'min_duration': 4,
+                       'max_duration': 4}
+
+    rail, rail_map = make_simple_rail2()
+
+    env = RailEnv(width=25,
+                  height=30,
+                  rail_generator=rail_from_grid_transition_map(rail),
+                  schedule_generator=random_schedule_generator(seed=2),  # seed 12
+                  number_of_agents=1,
+                  random_seed=1,
+                  stochastic_data=stochastic_data,  # Malfunction data generator
+                  )
+    env.reset()
+    # reset to initialize agents_static
+    env.agents[0].speed_data['speed'] = 1. / 3.
+    env.agents_static[0].target = (0, 0)
+
+    env.reset(False, False, True)
+    # Force malfunction to be off at beginning and next malfunction to happen in 2 steps
+    env.agents[0].malfunction_data['next_malfunction'] = 2
+    env.agents[0].malfunction_data['malfunction'] = 0
+    env_data = []
+    for step in range(20):
+        action_dict: Dict[int, RailEnvActions] = {}
+        for agent in env.agents:
+            # Go forward all the time
+            action_dict[agent.handle] = RailEnvActions(2)
+
+        if env.agents[0].malfunction_data['malfunction'] < 1:
+            agent_can_move = True
+        # Store the position before and after the step
+        pre_position = env.agents[0].speed_data['position_fraction']
+        _, reward, _, _ = env.step(action_dict)
+        # Check if the agent is still allowed to move in this step
+
+        if env.agents[0].malfunction_data['malfunction'] > 0:
+            agent_can_move = False
+        post_position = env.agents[0].speed_data['position_fraction']
+        # Assert that the agent moved while it was still allowed
+        if agent_can_move:
+            assert pre_position != post_position
+        else:
+            assert post_position == pre_position